tender-service-backend/pyproject.toml
[project]
name = "tender-service-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles==25.1.0",
    "alembic==1.16.5",
    "APScheduler==3.11.1",
    "async-lru==2.0.5",
    "asyncpg==0.31.0",
    "bcrypt==5.0.0",
    "fastapi[all]==0.125.0",
    "loguru==0.7.3",
    "openpyxl==3.1.5",
    "pandas==2.3.3",
    "Pillow==11.3.0",
    "psutil==7.1.3",
    "pydantic-validation-decorator==0.1.5",
    "PyJWT[crypto]==2.10.1",
    "psycopg2==2.9.11",
    "redis==6.4.0",
    "SQLAlchemy[asyncio]==2.0.45",
    "sqlglot[rs]==28.5.0",
    "user-agents==2.2.0",
    "requests>=2.32.5",
    "curl-cffi>=0.14.0",
    "langchain>=1.2.1",
    "langchain-deepseek>=1.0.1",
    "parsel>=1.10.0",
    "pymupdf>=1.24.10",
]

[dependency-groups]
dev = [
    "ruff==0.14.10",
]
